Added External Product and CMUX

+++ pyFHE/mulfft.py
@@ -8,7 +8,7 @@ def TwistGen(N):
 
 def TwistFFT(a,twist):
     Ns2 = len(twist)
-    b = np.array(a)
+    b = np.array(a,dtype = np.double)
     return np.fft.fft(np.multiply((b[:Ns2]+1j*b[Ns2:2*Ns2]),twist))
 
 def TwistIFFT(a,twist):
